+++ scryer-manifest.pl
@@ -1,5 +1,5 @@
 name("pure_regex").
-version("0.1.0.dev6").
+version("0.1.0.dev7").
 main_file("src/pure_regex.pl").
 license(name("Unlicense"), path("./UNLICENSE")).
 dependencies([]).